Reworked TrustRegionMinimizer to support various subproblems.
