+++ src/Numerics/Optimization/ITrustRegionSubProblem.cs
@@ -1,0 +1,13 @@
+﻿using MathNet.Numerics.LinearAlgebra;
+
+namespace MathNet.Numerics.Optimization
+{
+    public interface ITrustRegionSubproblem
+    {
+        Vector<double> Pstep { get; }
+        double PredictedReduction { get; }
+        bool HitBoundary { get; }
+
+        void Solve(IObjectiveModel objective, double radius);
+    }
+}
