+++ src/Numerics/Optimization/TrustRegionSubProblem.cs
@@ -1,0 +1,12 @@
+﻿using MathNet.Numerics.Optimization.Subproblems;
+
+namespace MathNet.Numerics.Optimization
+{
+    public static class TrustRegionSubproblem
+    {
+        public static ITrustRegionSubproblem Quadratic()
+        {
+            return new QuadraticSubproblem();
+        }
+    }
+}
